+++ HackerRank/yarra tsetnoc/J.c
@@ -1,4 +1,4 @@
-// whats the mistake?
+// what is the mistake?
 #include <stdio.h>
 
 int main()
@@ -6,23 +6,23 @@ int main()
     int n;
     scanf("%d", &n);
     int MyArray[n];
-    for (int i = 0; i < n; i++)
+    for (long long int i = 0; i < n; i++)
     {
-        scanf("%d", &MyArray[i]);
+        scanf("%lld", &MyArray[i]);
     }
-    int minn = MyArray[0];
-    for (int i = 0; i < n; i++)
+    long long int minn = MyArray[0];
+    for (long long int i = 0; i < n; i++)
     {
         if (minn > MyArray[i])
         {
             minn = MyArray[i];
         }
     }
-    for (int i = 0; i < n; i++)
+    for (long long int i = 0; i < n; i++)
     {
         if (MyArray[i] == minn)
         {
-            printf("%d", i);
+            printf("%lld", i);
             break;
         }
     }
